Fix library name on non-Windows platforms.
The Python sysconfig vars have the full file name of the Python library.
But the link flag can't have the prefix and suffix. Although many build
system deal with the full name, exporting to raw like generators, like
pkgconfig, causes problems. This change removes the prefix and suffix
as generically as possible to deal with both static and dynamic Python
lib names.